feat/set_and_sh: job handle notify

+++ srcs/job_control/jobexec/jobexec_fork_exec_interactive_job_sig_wrapper.c
@@ -23,13 +23,27 @@ t_lst               *data_tmp_jobs(void)
 }
 
 
-static void         push_job(t_job *job)
+static void         handle_job(int si_code, t_job *job)
 {
-    if(tcsetpgrp(0, getpid()) < 0)
-        LOG_ERROR("tcsetpgrp: %s", strerror(errno));
-    if(kill(job_get_kill_pid(job), SIGTSTP) < 0)
-        LOG_ERROR("kill: %s", strerror(errno));
-    job_mgr_env_push(job);
+    if (si_code == CLD_STOPPED)
+    {
+        if(tcsetpgrp(0, getpid()) < 0)
+            LOG_ERROR("tcsetpgrp: %s", strerror(errno));
+        if(kill(job_get_kill_pid(job), SIGTSTP) < 0)
+            LOG_ERROR("kill: %s", strerror(errno));
+        job_mgr_env_push(job);
+    }
+    if (job_mgr_find_by_pid(shenv_singleton()->jobs, job->pid)
+        && shenv_shflag_enabled(shenv_singleton(), "notify"))
+    {
+        if (si_code == CLD_EXITED || si_code == CLD_STOPPED
+            || si_code == CLD_KILLED)
+        {
+            LOG_INFO("si_code=CLD_EXITED, job->id=%d", job->pid);
+            job_print(job, 0);
+            job_mgr_remove(shenv_singleton()->jobs, job);
+        }
+    }
 }
 
 
@@ -40,28 +54,25 @@ static void			sigstp_catcher(int signum, siginfo_t *info, void *vp)
 
     gid = getpgid(info->si_pid);
     LOG_INFO("signum: %d, Signal %d from PID %d, code: %d, value: %d: gid: %d",
-    	signum, info->si_signo, (int)info->si_pid, info->si_code, info->si_value, gid);
-    if (info->si_code == CLD_STOPPED)
-	{
-        LOG_INFO("child stopped: %d", info->si_pid);
-        LOG_INFO("data_tmp_jobs() len: %d", twl_lst_len(data_tmp_jobs()));
-        job = job_mgr_find_by_pid(data_tmp_jobs(), info->si_pid);
-        if (!job)
-        {
-            LOG_INFO("search in shenv->jobs");
-            job = job_mgr_find_by_pid(shenv_singleton()->jobs, info->si_pid);
-        }
-        if (job)
-        {
-            LOG_DEBUG("info->si_pid: %d", info->si_pid);
-            LOG_DEBUG("job->pid: %d", job->pid);
-            push_job(job);
-        }
-        else
-        {
-            LOG_INFO("pid not found: %d", info->si_pid);
-        }
-	}
+	signum, info->si_signo, (int)info->si_pid, info->si_code, info->si_value, gid);
+    LOG_INFO("child stopped: %d", info->si_pid);
+    LOG_INFO("data_tmp_jobs() len: %d", twl_lst_len(data_tmp_jobs()));
+    job = job_mgr_find_by_pid(data_tmp_jobs(), info->si_pid);
+    if (!job)
+    {
+        LOG_INFO("search in shenv->jobs");
+        job = job_mgr_find_by_pid(shenv_singleton()->jobs, info->si_pid);
+    }
+    if (job)
+    {
+        LOG_DEBUG("info->si_pid: %d", info->si_pid);
+        LOG_DEBUG("job->pid: %d", job->pid);
+        handle_job(info->si_code, job);
+    }
+    else
+    {
+        LOG_INFO("pid not found: %d", info->si_pid);
+    }
     (void)vp;
 }
 
@@ -93,6 +104,6 @@ void				jobexec_fork_exec_interactive_job_sig_wrapper(t_job *job, void *ctx,
     job_mgr_pop(data_tmp_jobs(), job);
     LOG_INFO("tmp jobs len: %d: job->pid: %d (after exec)",
         twl_lst_len(data_tmp_jobs()), job->pid);
-    if (!job_mgr_find_by_pid(shenv_singleton()->jobs, job->pid))
+    if (job_mgr_find_by_pid(data_tmp_jobs(), job->pid))
         job_del(job);
 }
